Move new classes to appropriate packages
