+++ cli/src/main/java/com/devonfw/tools/ide/npm/NpmJsonVersions.java
@@ -1,4 +1,4 @@
-package com.devonfw.tools.ide.tool.npm;
+package com.devonfw.tools.ide.npm;
 
 import com.devonfw.tools.ide.common.JsonObject;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
